requirements.txt
# --- Web framework + API ---
Django==5.2.5                   # LTS (supports Python 3.10–3.13)
djangorestframework==3.16.1     # Works with Django 5.2
django-cors-headers==4.7.0
django-filter==25.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1

# --- App server + static files ---
gunicorn==23.0.0
whitenoise==6.9.0

# --- Auth (JWT) ---
djangorestframework-simplejwt==5.5.1

# --- Database (Postgres) ---
psycopg2-binary==2.9.10

# --- HTTP + utils ---
requests==2.32.3
urllib3==2.2.3
Markdown==3.6
Pillow==11.3.0

# --- CV / ML stack ---
opencv-contrib-python==4.11.0.86
numpy==2.3.2
scikit-image==0.25.2
pandas==2.3.1
matplotlib==3.10.5
scikit-learn==1.5.2
tensorflow==2.20.0
ipython==8.27.0
seaborn==0.13.2
supabase
python-dotenv
dj-database-url
